int => golang int32

+++ oz-seccomp/render_futex.go
@@ -50,17 +50,17 @@ func render_futex(pid int, args RegisterArgs) (string, error) {
 	// TODO: lots
 
 	if op == FUTEX_WAIT {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, args[2])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, int32(args[2]))
 	} else if op == FUTEX_WAKE {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, args[2])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, int32(args[2]))
 	} else if op == FUTEX_FD {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, args[2])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d)", uintptr(args[0]), opstr, int32(args[2]))
 	} else if op == FUTEX_WAKE_BITSET {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d, %d)", uintptr(args[0]), opstr, args[2], args[5])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d, %d)", uintptr(args[0]), opstr, int32(args[2]), args[5])
 	} else if op == FUTEX_WAIT_BITSET {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d, %d, %d", uintptr(args[0]), opstr, args[2], args[3], args[5])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d, %d, %d", uintptr(args[0]), opstr, int32(args[2]), args[3], args[5])
 	} else {
-		callrep = fmt.Sprintf("futex(0x%X, %s, %d, 0x%X, 0x%x, %d)", uintptr(args[0]), opstr, args[2], uintptr(args[3]), args[4], args[5])
+		callrep = fmt.Sprintf("futex(0x%X, %s, %d, 0x%X, 0x%x, %d)", uintptr(args[0]), opstr, int32(args[2]), uintptr(args[3]), args[4], args[5])
 	}
 	return callrep, nil
